+++ 404.php
@@ -10,7 +10,7 @@
           <label for="wp-block-search__input-1" class="wp-block-search__label screen-reader-text">Search</label>
           <div class="wp-block-search__inside-wrapper ">
             <input type="search" id="wp-block-search__input-1" class="wp-block-search__input " name="s" value="<?php echo esc_attr(get_search_query()); ?>" placeholder="" required="">
-            <button type="submit" class="wp-block-search__button  ">Search</button>
+            <button type="submit" class="wp-block-search__button wp-element-button">Search</button>
           </div>
         </form>
         <p>If that doesn't work, maybe you can find something else on the <a href="<?php echo esc_url(home_url()); ?>">front page</a>.</p>