+++ assets/js/block-styles.js
@@ -1,0 +1,4 @@
+wp.domReady(() => {
+  wp.blocks.unregisterBlockStyle('core/button', 'fill');
+	wp.blocks.unregisterBlockStyle('core/button', 'outline');
+});