+++ comments.php
@@ -33,7 +33,7 @@
       'comment_field'       => $comment_field,
       'comment_notes_before' => '',
       'logged_in_as'        => '<p class="logged-in-as">' . __('Logged in as', 'cpt-theme') . ' ' . wp_get_current_user()->display_name . '. <a href="' . wp_logout_url(get_permalink()) . '">' . __('Log out?', 'cpt-theme') . '</a></p>',
-      'submit_button'       => '<input name="%1$s" type="submit" id="%2$s" class="button %3$s" value="%4$s" />',
+      'submit_button'       => '<input name="%1$s" type="submit" id="%2$s" class="button wp-element-button %3$s" value="%4$s" />',
     ]);
   ?>
 </div><!--#comments-->