+++ theme-setup.php
@@ -18,7 +18,6 @@ function theme_setup() {
 	add_theme_support('post-thumbnails');
 	add_theme_support('responsive-embeds');
 	add_theme_support('title-tag');
-	add_theme_support('wp-block-styles');
 
   add_post_type_support('page', 'excerpt');